Rename typed experiment loader
