+++ isaaclab_arena/hydra/typed_experiment_loader.py
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-"""Hydra composition for typed Arena Experiments."""
+"""Load typed Arena Experiments through Hydra."""
 
 from __future__ import annotations
 
@@ -60,10 +60,10 @@ def load_arena_experiment_from_yaml(
     Returns:
         The typed Experiment Definition, preserving YAML mapping declaration order.
     """
-    run_values_by_name = load_arena_experiment_run_values_from_yaml(yaml_path)
+    run_values_by_name = load_experiment_run_definitions_from_yaml(yaml_path)
     config_store = ConfigStore.instance()
     # Reuse these internal names so repeated loads replace their process-global ConfigStore entries.
-    hydra_config_namespace = "isaaclab_arena_experiment_composition"
+    hydra_config_namespace = "isaaclab_arena_typed_experiment_loader"
 
     try:
         with _get_new_hydra_context_if_none_exists():
@@ -101,7 +101,7 @@ def _assert_run_name_is_hydra_compatible(run_name: str) -> None:
     )
 
 
-def load_arena_experiment_run_values_from_yaml(yaml_path: str | Path) -> dict[str, dict[str, Any]]:
+def load_experiment_run_definitions_from_yaml(yaml_path: str | Path) -> dict[str, dict[str, Any]]:
     """Read an Arena Experiment YAML file and return its Run values by name.
 
     This validates the shared YAML envelope. Fields belonging to a Run,
